feat: add shared proficiency utilities
